+++ mdadash/backend/analyses/dssp.py
@@ -49,6 +49,16 @@ class DSSPAnalysis(WidgetBase):
         Max values to show in plot
             Default: ``100``
 
+    Plot refresh frequency
+        The frequency with which the plot is refreshed (every n frames).
+        This only applies when the run frequency is ``every-frame``
+
+            Default: ``1``
+
+    Reset on connect
+        Reset the plot on every connect
+            Default: ``False``
+
     X-axis
         X-axis value - `time` or `step`
             Default: ``time``
@@ -107,6 +117,18 @@ class DSSPAnalysis(WidgetBase):
             "type": "int",
         },
         {
+            "attribute": "plot_refresh_frequency",
+            "name": "Plot refresh frequency",
+            "description": "The frequency with which the plot is refreshed (every n frames)",
+            "type": "int",
+        },
+        {
+            "attribute": "reset_on_connect",
+            "name": "Reset on connect",
+            "description": "Reset the plot on every connect",
+            "type": "bool",
+        },
+        {
             "attribute": "x_type",
             "name": "X-axis",
             "type": "toggle",
@@ -125,6 +147,9 @@ class DSSPAnalysis(WidgetBase):
         self.default_maxlen = 100
         self.maxlen = self.default_maxlen
         self.custom_title = None
+        self.plot_refresh_count = 1
+        self.plot_refresh_frequency = 1
+        self.reset_on_connect = False
         self.x_type = "time"
         self.x_values = None
         self._setup_plot()
@@ -158,6 +183,7 @@ class DSSPAnalysis(WidgetBase):
         self.steps = deque(maxlen=self.maxlen)
         self.times = deque(maxlen=self.maxlen)
         self.y_values = deque(maxlen=self.maxlen)
+        self.plot_refresh_count = 1
         self._set_x_values()
 
     def _set_title(self):
@@ -200,6 +226,9 @@ class DSSPAnalysis(WidgetBase):
         tick_spacing = max(1, self.n_residues // 10)
         self.ax.set_yticks(np.arange(0, self.n_residues, tick_spacing))
         self.ax.set_yticklabels(res_ids[::tick_spacing])
+        self.plot_refresh_count = 1
+        if self.reset_on_connect:  # pragma: no cover
+            self._reset_plot_values()
 
     def on_input_change(self, attribute, _old_value, new_value):
         """:meth:`~mdadash.backend.widgets.base.WidgetBase.on_input_change` handler"""
@@ -211,6 +240,8 @@ class DSSPAnalysis(WidgetBase):
             self._set_x_values()
         elif attribute == "custom_title":
             self._set_title()
+        elif attribute == "plot_refresh_frequency":
+            self.plot_refresh_count = 1
 
     def _compute_current_frame(self):
         """Compute values for current frame"""
@@ -257,16 +288,21 @@ class DSSPAnalysis(WidgetBase):
             self.steps.append(step)
             self.times.append(time)
         # update plot
-        matrix_to_plot = np.array(self.y_values).T
-        min_x = self.x_values[0]
-        max_x = self.x_values[-1]
-        if min_x == max_x:
-            max_x = min_x + 1
-        self.im.set_data(matrix_to_plot)
-        self.im.set_extent([min_x, max_x, 0, self.n_residues])
-        self.ax.relim()
-        self.ax.autoscale_view()
-        self.display_canvas(self.canvas)
+        if self._run_frequency == "batch" or (
+            self.plot_refresh_count == 1
+            or self.plot_refresh_count % self.plot_refresh_frequency == 0
+        ):
+            matrix_to_plot = np.array(self.y_values).T
+            min_x = self.x_values[0]
+            max_x = self.x_values[-1]
+            if min_x == max_x:
+                max_x = min_x + 1
+            self.im.set_data(matrix_to_plot)
+            self.im.set_extent([min_x, max_x, 0, self.n_residues])
+            self.ax.relim()
+            self.ax.autoscale_view()
+            self.display_canvas(self.canvas)
+        self.plot_refresh_count += 1
 
     def run_every_frame(self):
         """:meth:`~mdadash.backend.widgets.base.WidgetBase.run_every_frame` handler"""