+++ mdadash/backend/analyses/rmsd.py
@@ -70,6 +70,16 @@ class RMSD(WidgetBase):
         Max values to show in plot
             Default: ``100``
 
+    Plot refresh frequency
+        The frequency with which the plot is refreshed (every n frames).
+        This only applies when the run frequency is ``every-frame``
+
+            Default: ``1``
+
+    Reset on connect
+        Reset the plot on every connect
+            Default: ``False``
+
     X-axis
         X-axis value - `time` or `step`
             Default: ``time``
@@ -160,6 +170,18 @@ class RMSD(WidgetBase):
             "type": "int",
         },
         {
+            "attribute": "plot_refresh_frequency",
+            "name": "Plot refresh frequency",
+            "description": "The frequency with which the plot is refreshed (every n frames)",
+            "type": "int",
+        },
+        {
+            "attribute": "reset_on_connect",
+            "name": "Reset on connect",
+            "description": "Reset the plot on every connect",
+            "type": "bool",
+        },
+        {
             "attribute": "x_type",
             "name": "X-axis",
             "type": "toggle",
@@ -181,6 +203,9 @@ class RMSD(WidgetBase):
         self.custom_title = None
         self.default_maxlen = 100
         self.maxlen = self.default_maxlen
+        self.plot_refresh_count = 1
+        self.plot_refresh_frequency = 1
+        self.reset_on_connect = False
         self.x_type = "time"
         self.x_values = None
         self._setup_plot()
@@ -200,6 +225,7 @@ class RMSD(WidgetBase):
         self.steps = deque(maxlen=self.maxlen)
         self.times = deque(maxlen=self.maxlen)
         self.y_values = deque(maxlen=self.maxlen)
+        self.plot_refresh_count = 1
         self._set_x_values()
 
     def _set_title(self):
@@ -238,6 +264,9 @@ class RMSD(WidgetBase):
     def on_post_connect(self):
         """:meth:`~mdadash.backend.widgets.base.WidgetBase.on_post_connect` handler"""
         self._update_selection()
+        self.plot_refresh_count = 1
+        if self.reset_on_connect:  # pragma: no cover
+            self._reset_plot_values()
 
     def on_input_change(self, attribute, _old_value, new_value):
         """:meth:`~mdadash.backend.widgets.base.WidgetBase.on_input_change` handler"""
@@ -252,6 +281,8 @@ class RMSD(WidgetBase):
         elif attribute in ("selection", "center", "superposition"):
             self._reset_plot_values()
             self._update_selection()
+        elif attribute == "plot_refresh_frequency":
+            self.plot_refresh_count = 1
 
     def _compute_current_frame(self):
         """Compute values for current frame"""
@@ -286,10 +317,15 @@ class RMSD(WidgetBase):
             self.times.append(times)
             self.y_values.append(v)
         # update plot
-        self.plot.set_data(self.x_values, self.y_values)
-        self.ax.relim()
-        self.ax.autoscale_view()
-        self.display_canvas(self.canvas)
+        if self._run_frequency == "batch" or (
+            self.plot_refresh_count == 1
+            or self.plot_refresh_count % self.plot_refresh_frequency == 0
+        ):
+            self.plot.set_data(self.x_values, self.y_values)
+            self.ax.relim()
+            self.ax.autoscale_view()
+            self.display_canvas(self.canvas)
+        self.plot_refresh_count += 1
 
     def run_every_frame(self):
         """:meth:`~mdadash.backend.widgets.base.WidgetBase.run_every_frame` handler"""