+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The rules for this file:
 
 - Speed up of matplotlib plot generation (PR #82)
 - Use matplotlib Agg backend for built-in plots (PR #83)
+- Added refresh frequency and reset on connect options for plots (PR #84)
 
 ### Fixed